requirements.txt
graphiti-core
zep-python
pandas
python-dotenv
fastapi
uvicorn
spacy
graphiti